Added native integration of Kendra with Lex
